Check if tools are already installed before installing

+++ install-tools.sh
@@ -4,16 +4,23 @@
 # Ubuntu
 if [ -f /etc/os-release ] && [ "$(grep -c "Ubuntu" /etc/os-release)" -ne 0 ]; then
 
-    # Source: https://lindevs.com/install-exa-on-ubuntu
-    sudo apt update
-    sudo apt install -y unzip
-    EXA_VERSION=$(curl -s "https://api.github.com/repos/ogham/exa/releases/latest" | grep -Po '"tag_name": "v\K[0-9.]+')
-    curl -Lo exa.zip "https://github.com/ogham/exa/releases/latest/download/exa-linux-x86_64-v${EXA_VERSION}.zip"
-    sudo unzip -q exa.zip bin/exa -d /usr/local
-    rm -rf exa.zip
+    # exa (nice ls)
+    if ! command -v exa > /dev/null 2>&1; then
+        # Source: https://lindevs.com/install-exa-on-ubuntu
+        sudo apt update
+        sudo apt install -y unzip
+        EXA_VERSION=$(curl -s "https://api.github.com/repos/ogham/exa/releases/latest" | grep -Po '"tag_name": "v\K[0-9.]+')
+        curl -Lo exa.zip "https://github.com/ogham/exa/releases/latest/download/exa-linux-x86_64-v${EXA_VERSION}.zip"
+        sudo unzip -q exa.zip bin/exa -d /usr/local
+        rm -rf exa.zip
+    fi
 
-    # Source: https://github.com/wting/autojump
-    # OMZ plugin should help with making 'j' work
-    sudo apt-get install -y autojump
+    # autojump (nice cd)
+    if ! command -v autojump > /dev/null 2>&1; then
+        # Source: https://github.com/wting/autojump
+        # OMZ plugin should help with making 'j' work
+        sudo apt-get install -y autojump
+    fi
+
 
 fi
